# artifacts/api-server/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "0.1.0",
    "description": "Authenticated Stock Control and Ordering API. Send the configured API key as a Bearer token in the Authorization header."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base API path"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "operationId": "listOrders",
        "tags": [
          "orders"
        ],
        "summary": "List purchase orders",
        "description": "Returns purchase orders, optionally filtered by order number, a contained part number, reference, supplier or status.\n",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive partial match on Order Number."
          },
          {
            "name": "partNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive partial match on a line Part Number."
          },
          {
            "name": "reference",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Exact match on the order Reference (e.g. a Job Manager job id)."
          },
          {
            "name": "supplier",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive partial match on Supplier Name."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Draft",
                "Submitted",
                "Confirmed",
                "Received",
                "Cancelled"
              ]
            },
            "description": "Exact match on Status."
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase order summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderSummary"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "createOrder",
        "tags": [
          "orders"
        ],
        "summary": "Create a purchase order",
        "description": "If orderNumber is omitted the server generates the next PO-nnnn number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Purchase order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Order Number already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/summary": {
      "get": {
        "operationId": "getOrderSummary",
        "tags": [
          "orders"
        ],
        "summary": "Get order summary",
        "description": "Returns purchase order totals and status counts.",
        "responses": {
          "200": {
            "description": "Purchase order summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummaryTotals"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/order/{orderNumber}": {
      "get": {
        "operationId": "getOrder",
        "tags": [
          "orders"
        ],
        "summary": "Retrieve a purchase order",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Purchase order with lines",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateOrder",
        "tags": [
          "orders"
        ],
        "summary": "Update a purchase order",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderNumber"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated purchase order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteOrder",
        "tags": [
          "orders"
        ],
        "summary": "Delete a purchase order",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderNumber"
          }
        ],
        "responses": {
          "204": {
            "description": "Purchase order deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/order/{orderNumber}/lines": {
      "post": {
        "operationId": "appendOrderLine",
        "tags": [
          "orders"
        ],
        "summary": "Add a line to a Draft purchase order",
        "description": "Appends a line with the next line number. If the part is already on the order its quantity is increased instead. Only Draft orders can be changed (409 otherwise).\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/OrderNumber"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderLineAppend"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated purchase order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Order is not a Draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Use the configured STOCK_API_KEY (or SESSION_SECRET in development) as the bearer token."
      }
    },
    "parameters": {
...
      "OrderNumber": {
        "name": "orderNumber",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "The unique Order Number."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Stock item not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
...
      "OrderLine": {
        "type": "object",
        "properties": {
          "lineNumber": {
            "type": "integer",
            "minimum": 1
          },
          "partNumber": {
            "type": "string"
          },
          "externalPartNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "lineNumber",
          "partNumber",
          "externalPartNumber",
          "description",
          "quantity",
          "unitPrice"
        ]
      },
      "OrderLineInput": {
        "type": "object",
        "properties": {
          "lineNumber": {
            "type": "integer",
            "minimum": 1
          },
          "partNumber": {
            "type": "string",
            "minLength": 1
          },
          "externalPartNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "lineNumber",
          "partNumber",
          "externalPartNumber",
          "description",
          "quantity",
          "unitPrice"
        ]
      },
      "OrderLineAppend": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "string",
            "minLength": 1
          },
          "externalPartNumber": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "partNumber",
          "quantity",
          "unitPrice"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "supplierName": {
            "type": "string"
          },
          "reference": {
            "type": "string",
            "description": "Free-text reference to the originating record, e.g. a Job Manager job id."
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Confirmed",
              "Received",
              "Cancelled"
            ]
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderLine"
            }
          },
          "totalValue": {
            "type": "number",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "reference",
          "orderNumber",
          "orderDate",
          "supplierName",
          "status",
          "lines",
          "totalValue",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderInput": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string",
            "minLength": 1
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "supplierName": {
            "type": "string",
            "minLength": 1
          },
          "reference": {
            "type": "string",
            "description": "Free-text reference to the originating record, e.g. a Job Manager job id."
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Confirmed",
              "Received",
              "Cancelled"
            ]
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          }
        },
        "required": [
          "orderDate",
          "supplierName",
          "status",
          "lines"
        ]
      },
      "OrderUpdate": {
        "type": "object",
        "properties": {
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "supplierName": {
            "type": "string",
            "minLength": 1
          },
          "reference": {
            "type": "string",
            "description": "Free-text reference to the originating record, e.g. a Job Manager job id."
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Confirmed",
              "Received",
              "Cancelled"
            ]
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/OrderLineInput"
            }
          }
        },
        "minProperties": 1
      },
      "OrderSummary": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string"
          },
          "orderDate": {
            "type": "string",
            "format": "date"
          },
          "supplierName": {
            "type": "string"
          },
          "reference": {
            "type": "string",
            "description": "Free-text reference to the originating record, e.g. a Job Manager job id."
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Submitted",
              "Confirmed",
              "Received",
              "Cancelled"
            ]
          },
          "lineCount": {
            "type": "integer"
          },
          "totalValue": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "reference",
          "orderNumber",
          "orderDate",
          "supplierName",
          "status",
          "lineCount",
          "totalValue"
        ]
      },
      "OrderSummaryTotals": {
        "type": "object",
        "properties": {
          "orderCount": {
            "type": "integer"
          },
          "openOrderCount": {
            "type": "integer"
          },
          "totalValue": {
            "type": "number"
          },
          "statusCounts": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        },
        "required": [
          "orderCount",
          "openOrderCount",
          "totalValue",
          "statusCounts"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}