# lib/jobs-api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Authenticated Job Manager API. Send the configured API key as a Bearer token in the Authorization header. Adding a part to a job takes stock from the Stock Control API and, for any shortfall, adds a line to a Draft order in the Ordering API.
 *
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  error: 'error',
} as const;

export type HealthStatusDatabase = typeof HealthStatusDatabase[keyof typeof HealthStatusDatabase];


export const HealthStatusDatabase = {
  ok: 'ok',
  error: 'error',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
  database: HealthStatusDatabase;
}

export interface Error {
  error: string;
}

export interface Part {
  partNumber: string;
  itemName: string;
  description: string;
  /** Quantity currently in stock. */
  quantity: number;
  binNumber: string;
}

export interface JobPart {
  partNumber: string;
  description: string;
  /**
     * Total quantity the job needs.
     * @minimum 1
     */
  quantity: number;
  /**
     * Quantity taken from stock.
     * @minimum 0
     */
  quantityAllocated: number;
  /**
     * Quantity placed on a Draft order because stock was short.
     * @minimum 0
     */
  quantityOrdered: number;
  /**
     * The Draft order holding the shortfall, if any.
     * @nullable
     */
  draftOrderNumber: string | null;
}

export interface JobPartInput {
  /** @minLength 1 */
  partNumber: string;
  /** @minimum 1 */
  quantity: number;
}

export interface Job {
  jobId: string;
  jobDate: string;
  client: string;
  clientAddress: string;
  description: string;
  parts: JobPart[];
  createdAt: string;
  updatedAt: string;
}

export interface JobInput {
  jobDate: string;
  /** @minLength 1 */
  client: string;
  clientAddress: string;
  description: string;
}

export interface JobUpdate {
  jobDate?: string;
  /** @minLength 1 */
  client?: string;
  clientAddress?: string;
  description?: string;
}

export interface JobSummary {
  jobId: string;
  jobDate: string;
  client: string;
  description: string;
  partCount: number;
}

export interface JobSummaryTotals {
  /** Total number of jobs. */
  jobCount: number;
  /** Total number of part lines across all jobs. */
  partCount: number;
  /** Sum of quantity across all job parts. */
  totalQuantity: number;
  /** Sum of quantity * Stock Control cost across all job parts. Parts no longer in stock contribute 0. */
  totalValue: number;
}

/**
 * Missing or invalid API key
 */
export type UnauthorizedResponse = Error;

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Not found
 */
export type NotFoundResponse = Error;

/**
 * The Stock Control or Ordering API failed or could not be reached
 */
export type UpstreamFailureResponse = Error;

export type ListPartsParams = {
/**
 * Case-insensitive partial match on Part Number.
 */
search?: string;
};

export type ListJobsParams = {
/**
 * Case-insensitive partial match on Job Id, Client or a Part Number on the job.
 */
search?: string;
};